The game is finished. An intro story and an ending story have been added to the game. A game-ending menu has also been implemented.
I have fully completed the game, and this is its final version. I added the game's story to the intro screen. I also added a game-ending menu to the in-game menu. After the game ends, I included an ending story, and the player can choose to continue playing or end the game.